Added unit additional unit tests for login, signup and getMe
Added unit tests that cover true negtaive cases, such as incorrect
credentials when logging in or trying to register a user with an
existing username. Also reformatted the status codes into constants in
the init file in the user app to provide uniformity across all files
and prevent unwanted bugs.