Fixed a bug regarding collection creation
The code only checked for the name of the collection so when a user
tried to createa a collection with the same name of another collection
(regardless if created by the user or not) it would not allow it.

Now the code checks for the name of the collection and the owner.
Also reformatted the name of the variable that contains the collection
for better clarity